Refactor user context
