--- app/models/user_contexts/admin_user_context.rb
@@ -1,15 +1,0 @@
-module UserContexts
-	class AdminUserContext < UserContext
-		def reports
-      Report.all
-	  end
-
-		def phds_list
-	    PHD.all.pluck(:name, :id)
-	  end
-
-	  def ods_list(place_id)
-    	place_id.present? ? Place.find(place_id).children.pluck(:name, :id) : []
-	  end
-	end
-end
